// ProposalToFigma/Codex/proposal-to-figma/apps/figma-plugin/src/shared/types.ts
export type OutputType =
  | "BANNER"
  | "STORY"
  | "SNS_X_OPEN_01"
  | "SNS_X_OPEN_02"
  | "SNS_X_OPEN_03"
  | "SNS_X_UPDATE";
// ...
export type TemplateLocator = {
  outputType: OutputType;
  sectionName: string;
  templateLabel: string;
  templateInstanceName: string;
};